+++ crates/matrix-sdk-ui/tests/integration/timeline/sliding_sync.rs
@@ -24,7 +24,9 @@ use matrix_sdk::{
     test_utils::logged_in_client_with_server,
 };
 use matrix_sdk_test::{async_test, mocks::mock_encryption_state};
-use matrix_sdk_ui::timeline::{TimelineBuilder, TimelineItem, TimelineItemKind};
+use matrix_sdk_ui::timeline::{
+    TimelineBuilder, TimelineItem, TimelineItemKind, TimelineReadReceiptTracking,
+};
 use ruma::{RoomId, room_id, user_id};
 use serde_json::json;
 use wiremock::{Match, Mock, MockServer, Request, ResponseTemplate, http::Method};
@@ -404,7 +406,10 @@ async fn timeline_test_helper(
         anyhow::anyhow!("Room {room_id} not found in client. Can't provide a timeline for it")
     })?;
 
-    let timeline = TimelineBuilder::new(&sdk_room).track_read_marker_and_receipts().build().await?;
+    let timeline = TimelineBuilder::new(&sdk_room)
+        .track_read_marker_and_receipts(TimelineReadReceiptTracking::AllEvents)
+        .build()
+        .await?;
 
     Ok(timeline.subscribe().await)
 }